Copy Artifacts from S3
This new extension allows to copy artifacts uploaded to S3 by a previous
run to the current build workspace, like the "Copy Artifacts" plugin
does.

It also fingerprint downloaded artifacts to be able to link builds
together.

This extension requires the Copy Artifact plugin to be installed to work
correctly.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>